Split jenkins settings and jenkins triggers into two classes
