+++ pom.xml
@@ -174,6 +174,17 @@
             <artifactId>byte-buddy</artifactId>
             <version>1.10.2</version>
             <scope>test</scope>
+        <dependency>
+            <groupId>org.powermock</groupId>
+            <artifactId>powermock-module-junit4</artifactId>
+            <version>${powermock.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.powermock</groupId>
+            <artifactId>powermock-api-mockito2</artifactId>
+            <version>${powermock.version}</version>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -338,5 +349,6 @@
         <atlassian.spring.scanner.version>2.2.0</atlassian.spring.scanner.version>
         <amps.version>8.0.2</amps.version>
         <plugin.testrunner.version>2.0.1</plugin.testrunner.version>
+        <powermock.version>2.0.2</powermock.version>
     </properties>
 </project>
