+++ src/main/java/com/kylenicholls/stash/parameterizedbuilds/ciserver/AccountServer.java
@@ -21,6 +21,7 @@ public class AccountServer extends CIServer {
 
     public AccountServer(Jenkins jenkins, ApplicationUser user, ProjectService projectService){
         this.jenkins = jenkins;
+        this.jenkinsConn = new JenkinsConnection(jenkins);
         this.user = user;
         this.projectService = projectService;
         this.JENKINS_SETTINGS = "jenkins.user.settings";
