+++ src/main/java/com/kylenicholls/stash/parameterizedbuilds/eventHandlers/PushHandler.java
@@ -49,7 +49,7 @@ public class PushHandler extends RefHandler {
 
                 private boolean triggerJob(Change change) {
                     if (change.getPath().toString().matches(pathRegex)) {
-                        jenkins.triggerJob(projectKey, user, job, bitbucketVariables);
+                        jenkinsConn.triggerJob(projectKey, user, job, bitbucketVariables);
                         return false;
                     }
                     return true;
