+++ src/main/java/com/kylenicholls/stash/parameterizedbuilds/rest/ProjectResource.java
@@ -25,6 +25,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 import com.kylenicholls.stash.parameterizedbuilds.ciserver.Jenkins;
+import com.kylenicholls.stash.parameterizedbuilds.ciserver.JenkinsConnection;
 import com.kylenicholls.stash.parameterizedbuilds.item.Server;
 import com.sun.jersey.spi.resource.Singleton;
 
@@ -71,7 +72,8 @@ public class ProjectResource extends RestResource implements ServerService {
             Server oldServer = jenkins.getJenkinsServer(projectKey);
             server.setToken(getCurrentDefaultToken(oldServer, server));
 
-            String message = jenkins.testConnection(server);
+            JenkinsConnection jenkinsConn = new JenkinsConnection(jenkins);
+            String message = jenkinsConn.testConnection(server);
 
             if(message.equals("Connection successful")){
                 return Response.ok(message).build();
