+++ src/test/java/com/kylenicholls/stash/parameterizedbuilds/PullRequestHookTest.java
@@ -16,6 +16,10 @@ import com.atlassian.bitbucket.hook.repository.RepositoryHook;
 import com.kylenicholls.stash.parameterizedbuilds.eventHandlers.TestEventFactory;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.atlassian.bitbucket.project.Project;
 import com.atlassian.bitbucket.pull.PullRequestService;
@@ -23,15 +27,17 @@ import com.atlassian.bitbucket.repository.Repository;
 import com.atlassian.bitbucket.server.ApplicationPropertiesService;
 import com.atlassian.bitbucket.setting.Settings;
 import com.kylenicholls.stash.parameterizedbuilds.ciserver.Jenkins;
+import com.kylenicholls.stash.parameterizedbuilds.ciserver.JenkinsConnection;
 import com.kylenicholls.stash.parameterizedbuilds.helper.SettingsService;
 import com.kylenicholls.stash.parameterizedbuilds.item.Job;
 import com.kylenicholls.stash.parameterizedbuilds.item.Job.JobBuilder;
 import com.kylenicholls.stash.parameterizedbuilds.item.Server;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.concurrent.ExecutorService;
 
+@RunWith(PowerMockRunner.class)
+@PrepareForTest(PullRequestHook.class)
 public class PullRequestHookTest {
     private final String COMMIT = "commithash";
     private final String PR_URI = "http://pruri";
@@ -39,6 +45,7 @@ public class PullRequestHookTest {
             false, false);
     private SettingsService settingsService;
     private Jenkins jenkins;
+    private JenkinsConnection jenkinsConn;
     private ApplicationPropertiesService propertiesService;
     private PullRequestHook hook;
     private Repository repository;
@@ -49,7 +56,7 @@ public class PullRequestHookTest {
     private ExecutorService executorService;
 
     @Before
-    public void setup() throws URISyntaxException {
+    public void setup() throws Exception {
         settingsService = mock(SettingsService.class);
         PullRequestService pullRequestService = mock(PullRequestService.class);
         jenkins = mock(Jenkins.class);
@@ -84,6 +91,11 @@ public class PullRequestHookTest {
                 .pathRegex("").prDestRegex("");
         jobs = new ArrayList<>();
         when(settingsService.getJobs(any())).thenReturn(jobs);
+
+        jenkinsConn = mock(JenkinsConnection.class);
+        PowerMockito.whenNew(JenkinsConnection.class)
+            .withArguments(jenkins)
+            .thenReturn(jenkinsConn);
     }
 
     @Test
@@ -94,8 +106,7 @@ public class PullRequestHookTest {
         when(rescopedEvent.getPreviousFromHash()).thenReturn("newhash");
         hook.onPullRequestRescoped(rescopedEvent);
 
-        verify(jenkins, times(1))
-                .triggerJob(any(), any(), any(), any());
+        verify(jenkinsConn, times(1)).triggerJob(any(), any(), any(), any());
     }
 
     @Test
@@ -106,6 +117,6 @@ public class PullRequestHookTest {
         when(rescopedEvent.getPreviousFromHash()).thenReturn(COMMIT);
         hook.onPullRequestRescoped(rescopedEvent);
 
-        verify(jenkins, times(1)).triggerJob(any(), any(), any(), any());
+        verify(jenkinsConn, times(1)).triggerJob(any(), any(), any(), any());
     }
 }
