+++ src/test/java/com/kylenicholls/stash/parameterizedbuilds/rest/BuildResourceTest.java
@@ -13,6 +13,10 @@ import com.atlassian.bitbucket.pull.PullRequestRef;
 import com.google.common.collect.Lists;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.atlassian.bitbucket.auth.AuthenticationContext;
 import com.atlassian.bitbucket.i18n.I18nService;
@@ -25,6 +29,7 @@ import com.atlassian.bitbucket.server.ApplicationPropertiesService;
 import com.atlassian.bitbucket.setting.Settings;
 import com.atlassian.bitbucket.user.ApplicationUser;
 import com.kylenicholls.stash.parameterizedbuilds.ciserver.Jenkins;
+import com.kylenicholls.stash.parameterizedbuilds.ciserver.JenkinsConnection;
 import com.kylenicholls.stash.parameterizedbuilds.conditions.BuildPermissionsCondition;
 import com.kylenicholls.stash.parameterizedbuilds.helper.SettingsService;
 import com.kylenicholls.stash.parameterizedbuilds.item.JenkinsResponse;
@@ -41,12 +46,15 @@ import java.util.Map;
 
 import static org.mockito.Mockito.when;
 
+@RunWith(PowerMockRunner.class)
+@PrepareForTest(BuildResource.class)
 public class BuildResourceTest {
     private final String REPO_SLUG = "reposlug";
     private final String PROJECT_KEY = "projkey";
     private final String URI = "http://uri";
     private BuildResource rest;
     private Jenkins jenkins;
+    private JenkinsConnection jenkinsConn;
     private Repository repository;
     private AuthenticationContext authContext;
     private SettingsService settingsService;
@@ -94,6 +102,11 @@ public class BuildResourceTest {
         when(settingsService.getJobs(any())).thenReturn(jobs);
         hook = mock(RepositoryHook.class);
         when(settingsService.getHook(any())).thenReturn(hook);
+        
+        jenkinsConn = mock(JenkinsConnection.class);
+        PowerMockito.whenNew(JenkinsConnection.class)
+            .withArguments(jenkins)
+            .thenReturn(jenkinsConn);
     }
 
     @Test
@@ -133,7 +146,7 @@ public class BuildResourceTest {
         query.add("param1", "value1");
         query.add("param2", "value2");
         when(uriInfo.getQueryParameters()).thenReturn(query);
-        when(jenkins.triggerJob(any(), any(), any(), any())).thenReturn(message);
+        when(jenkinsConn.triggerJob(any(), any(), any(), any())).thenReturn(message);
         Response results = rest.triggerBuild(repository, "0", "test", uriInfo);
 
         assertEquals(Response.Status.OK.getStatusCode(), results.getStatus());
@@ -230,7 +243,7 @@ public class BuildResourceTest {
         when(settingsService.getSettings(repository)).thenReturn(null);
         Response actual = rest.getJobs(repository, "branch", "commit", null, 0);
 
-        assertEquals(Response.Status.OK.getStatusCode(), actual.getStatus());
+        assertEquals(Response.Status.NOT_FOUND.getStatusCode(), actual.getStatus());
     }
 
     @SuppressWarnings("unchecked")
